split http/recv tests
