+++ gurl_test.go
@@ -9,7 +9,6 @@
 package gurl_test
 
 import (
-	"errors"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -223,152 +222,152 @@ func (c Seq) Value(i int) interface{} { return c[i] }
 // 	it.Ok(t).If(io.Fail).Should().Equal(nil)
 // }
 
-func TestCodeOk(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestCodeOk(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestCodeFail(t *testing.T) {
-	ts := httptest.NewServer(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(http.StatusNotFound)
-		}),
-	)
-	defer ts.Close()
+// func TestCodeFail(t *testing.T) {
+// 	ts := httptest.NewServer(
+// 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+// 			w.WriteHeader(http.StatusNotFound)
+// 		}),
+// 	)
+// 	defer ts.Close()
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ƒ.Code(200),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ƒ.Code(200),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).ShouldNot().Equal(nil).
-		If(io.Fail).Should().Equal(&gurl.BadMatchCode{[]int{200}, 404})
-}
+// 	it.Ok(t).
+// 		If(io.Fail).ShouldNot().Equal(nil).
+// 		If(io.Fail).Should().Equal(&gurl.BadMatchCode{[]int{200}, 404})
+// }
 
-func TestHeaderOk(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestHeaderOk(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.ServedJSON(),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.ServedJSON(),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestHeaderAny(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestHeaderAny(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Header("Content-Type").Any(),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Header("Content-Type").Any(),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestHeaderValue(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestHeaderValue(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var mime string
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Header("Content-Type").String(&mime),
-	)(gurl.IO())
+// 	var mime string
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Header("Content-Type").String(&mime),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().Equal(nil).
-		If(mime).Should().Equal("application/json")
-}
+// 	it.Ok(t).
+// 		If(io.Fail).Should().Equal(nil).
+// 		If(mime).Should().Equal("application/json")
+// }
 
-func TestHeaderFail(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestHeaderFail(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.ServedForm(),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.ServedForm(),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).ShouldNot().Equal(nil).
-		If(io.Fail).Should().Equal(&gurl.BadMatchHead{"Content-Type", "application/x-www-form-urlencoded", "application/json"})
-}
+// 	it.Ok(t).
+// 		If(io.Fail).ShouldNot().Equal(nil).
+// 		If(io.Fail).Should().Equal(&gurl.BadMatchHead{"Content-Type", "application/x-www-form-urlencoded", "application/json"})
+// }
 
-func TestRecvJSON(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestRecvJSON(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.ServedJSON(),
-		ƒ.Recv(&data),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.ServedJSON(),
+// 		ƒ.Recv(&data),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().Equal(nil).
-		If(data.Site).Should().Equal("example.com")
-}
+// 	it.Ok(t).
+// 		If(io.Fail).Should().Equal(nil).
+// 		If(data.Site).Should().Equal("example.com")
+// }
 
-func TestRecvForm(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestRecvForm(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptForm(),
-		ƒ.Code(200),
-		ƒ.ServedForm(),
-		ƒ.Recv(&data),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptForm(),
+// 		ƒ.Code(200),
+// 		ƒ.ServedForm(),
+// 		ƒ.Recv(&data),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().Equal(nil).
-		If(data.Site).Should().Equal("example.com")
+// 	it.Ok(t).
+// 		If(io.Fail).Should().Equal(nil).
+// 		If(data.Site).Should().Equal("example.com")
 
-}
+// }
 
-func TestBytes(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestBytes(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data []byte
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.ServedJSON(),
-		ƒ.Bytes(&data),
-	)(gurl.IO())
+// 	var data []byte
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.ServedJSON(),
+// 		ƒ.Bytes(&data),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().Equal(nil).
-		If(string(data)).Should().Equal("{\"site\": \"example.com\"}")
-}
+// 	it.Ok(t).
+// 		If(io.Fail).Should().Equal(nil).
+// 		If(string(data)).Should().Equal("{\"site\": \"example.com\"}")
+// }
 
 func TestJoin(t *testing.T) {
 	ts := mock()
@@ -387,175 +386,175 @@ func TestJoin(t *testing.T) {
 	it.Ok(t).If(io.Fail).Should().Equal(nil)
 }
 
-func TestDefined(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestDefined(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Defined(&data.Site),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Defined(&data.Site),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestNotDefined(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestNotDefined(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Defined(&data.Host),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Defined(&data.Host),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(&gurl.Undefined{"string"})
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(&gurl.Undefined{"string"})
+// }
 
-func TestValue(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestValue(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Value(&data).Is(&Test{Site: "example.com"}),
-		ƒ.Value(&data.Site).String("example.com"),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Value(&data).Is(&Test{Site: "example.com"}),
+// 		ƒ.Value(&data.Site).String("example.com"),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestRequireFail(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestRequireFail(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Value(&data.Site).String("localhost"),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Value(&data.Site).String("localhost"),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().Be().Like(&gurl.Mismatch{})
-}
+// 	it.Ok(t).
+// 		If(io.Fail).Should().Be().Like(&gurl.Mismatch{})
+// }
 
-func TestLookup(t *testing.T) {
-	ts := mockSeq()
-	defer ts.Close()
+// func TestLookup(t *testing.T) {
+// 	ts := mockSeq()
+// 	defer ts.Close()
 
-	var data Seq
-	expectS := Test{Site: "s.example.com"}
-	expectZ := Test{Site: "z.example.com"}
+// 	var data Seq
+// 	expectS := Test{Site: "s.example.com"}
+// 	expectZ := Test{Site: "z.example.com"}
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Seq(&data).Has(expectS.Site),
-		ƒ.Seq(&data).Has(expectS.Site, expectS),
-		ƒ.Seq(&data).Has(expectZ.Site, expectZ),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Seq(&data).Has(expectS.Site),
+// 		ƒ.Seq(&data).Has(expectS.Site, expectS),
+// 		ƒ.Seq(&data).Has(expectZ.Site, expectZ),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestLookupFailure(t *testing.T) {
-	ts := mockSeq()
-	defer ts.Close()
+// func TestLookupFailure(t *testing.T) {
+// 	ts := mockSeq()
+// 	defer ts.Close()
 
-	var data Seq
-	expect0 := Test{Site: "0.example.com"}
+// 	var data Seq
+// 	expect0 := Test{Site: "0.example.com"}
 
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Seq(&data).Has(expect0.Site),
-	)(gurl.IO())
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Seq(&data).Has(expect0.Site),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Be().Like(&gurl.Undefined{})
-}
+// 	it.Ok(t).If(io.Fail).Should().Be().Like(&gurl.Undefined{})
+// }
 
-func TestLookupMismatch(t *testing.T) {
-	ts := mockSeq()
-	defer ts.Close()
+// func TestLookupMismatch(t *testing.T) {
+// 	ts := mockSeq()
+// 	defer ts.Close()
 
-	var data Seq
-	expectS := Test{Site: "s.example.com"}
-	expectZ := Test{Site: "z.example.com"}
+// 	var data Seq
+// 	expectS := Test{Site: "s.example.com"}
+// 	expectZ := Test{Site: "z.example.com"}
 
-	err := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.Seq(&data).Has(expectS.Site, expectZ),
-	)(gurl.IO()).Fail
+// 	err := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.Seq(&data).Has(expectS.Site, expectZ),
+// 	)(gurl.IO()).Fail
 
-	it.Ok(t).
-		If(strings.Contains(err.Error(), `Site: "s.example.com"`)).Should().Equal(true).
-		If(strings.Contains(err.Error(), `Site: "z.example.com"`)).Should().Equal(true)
-}
+// 	it.Ok(t).
+// 		If(strings.Contains(err.Error(), `Site: "s.example.com"`)).Should().Equal(true).
+// 		If(strings.Contains(err.Error(), `Site: "z.example.com"`)).Should().Equal(true)
+// }
 
-func TestAssert(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestAssert(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.FMap(func() (err error) {
-			if data.Site != "example.com" {
-				err = errors.New("something wrong!")
-			}
-			return
-		}),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.FMap(func() (err error) {
+// 			if data.Site != "example.com" {
+// 				err = errors.New("something wrong!")
+// 			}
+// 			return
+// 		}),
+// 	)(gurl.IO())
 
-	it.Ok(t).If(io.Fail).Should().Equal(nil)
-}
+// 	it.Ok(t).If(io.Fail).Should().Equal(nil)
+// }
 
-func TestAssertFailure(t *testing.T) {
-	ts := mock()
-	defer ts.Close()
+// func TestAssertFailure(t *testing.T) {
+// 	ts := mock()
+// 	defer ts.Close()
 
-	var data Test
-	io := gurl.HTTP(
-		ø.GET(ts.URL),
-		ø.AcceptJSON(),
-		ƒ.Code(200),
-		ƒ.Recv(&data),
-		ƒ.FMap(func() (err error) {
-			if data.Site == "example.com" {
-				err = errors.New("something wrong!")
-			}
-			return
-		}),
-	)(gurl.IO())
+// 	var data Test
+// 	io := gurl.HTTP(
+// 		ø.GET(ts.URL),
+// 		ø.AcceptJSON(),
+// 		ƒ.Code(200),
+// 		ƒ.Recv(&data),
+// 		ƒ.FMap(func() (err error) {
+// 			if data.Site == "example.com" {
+// 				err = errors.New("something wrong!")
+// 			}
+// 			return
+// 		}),
+// 	)(gurl.IO())
 
-	it.Ok(t).
-		If(io.Fail).Should().
-		Equal(errors.New("something wrong!"))
-}
+// 	it.Ok(t).
+// 		If(io.Fail).Should().
+// 		Equal(errors.New("something wrong!"))
+// }
 
 func TestStatusSuccess(t *testing.T) {
 	ts := mock()
