+++ http/send/arrows_test.go
@@ -50,10 +50,22 @@ func TestMethod(t *testing.T) {
 func TestURL(t *testing.T) {
 	mthd := []func(string, ...interface{}) gurl.Arrow{ø.GET, ø.POST, ø.PUT, ø.DELETE}
 	for _, f := range mthd {
-		io := f("https://example.com/%v", 1)(gurl.IO())
+		io := f("https://example.com/%s/%v", "a", 1)(gurl.IO())
 		it.Ok(t).
 			If(io.Fail).Should().Equal(nil).
-			If(io.URL.String()).Should().Equal("https://example.com/1")
+			If(io.URL.String()).Should().Equal("https://example.com/a/1")
+	}
+}
+
+func TestURLByRef(t *testing.T) {
+	a := "a"
+	b := 1
+	mthd := []func(string, ...interface{}) gurl.Arrow{ø.GET, ø.POST, ø.PUT, ø.DELETE}
+	for _, f := range mthd {
+		io := f("https://example.com/%s/%v", &a, &b)(gurl.IO())
+		it.Ok(t).
+			If(io.Fail).Should().Equal(nil).
+			If(io.URL.String()).Should().Equal("https://example.com/a/1")
 	}
 }
